# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Port Daddy API
  version: 3.24.1
  description: >
    Authoritative port management and service orchestration for multi-agent
    development. Provides atomic port assignment, pub/sub messaging, distributed
    locks, agent registry, session coordination, and project:stack:context
    semantic identities.
  license:
    name: MIT
servers:
  - url: http://localhost:9876
    description: Local daemon default. Use `pd status` or `PORT_DADDY_URL` when your live daemon is elsewhere.
...
paths:
...
  /spawn:
    post:
      tags: [spawn]
      summary: Launch an AI agent
      operationId: spawnAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [backend, identity, budgetUsd, task]
              properties:
                backend:
                  type: string
                  enum: [ollama, claude, claude-cli, gemini, cloudflare, codex, aider, custom]
                model:
                  type: string
                  description: Model name (backend-specific)
                modelTier:
                  type: string
                  enum: [low, mid, high]
                  description: Optional backend-specific model tier shortcut
                identity:
                  type: string
                  description: Semantic identity for the spawned agent and project budget attribution
                budgetUsd:
                  type: number
                  description: Positive spend ceiling for this launch in USD
                purpose:
                  type: string
                task:
                  type: string
                  description: Task description or prompt
                files:
                  type: array
                  items:
                    type: string
                workdir:
                  type: string
                env:
                  type: object
                  additionalProperties:
                    type: string
                  description: Extra environment variables
                timeout:
                  type: integer
                  description: Timeout in milliseconds
      responses:
        '200':
          description: Agent spawned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  agentId:
                    type: string
                  backend:
                    type: string
                  status:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [spawn]
      summary: List active spawned agents
      operationId: listSpawnedAgents
      responses:
        '200':
          description: Active spawned agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  agents:
                    type: array
                    items: {}
                  count:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalError'
  /spawn/{id}:
    delete:
      tags: [spawn]
      summary: Kill a spawned agent
      operationId: killSpawnedAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent killed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  agentId:
                    type: string
                  message:
                    type: string
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # HARBORS
  # ===========================================================================
...
components:
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
                example: VALIDATION_ERROR
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: internal server error