# routes/spawn.ts
/**
 * Spawn Routes — AI Agent Launcher
 *
 * POST /spawn        — launch an AI agent, body: SpawnSpec, returns SpawnResult
 * GET  /spawn        — list active spawned agents
 * DELETE /spawn/:id  — kill a spawned agent
 */

import type { FastifyPluginAsync } from 'fastify';
import type { SpawnSpec, Spawner } from '../lib/spawner.js';
import { assessSpawnPreflight } from '../lib/spawn-preflight.js';
import type { CostTracker } from '../lib/cost-tracker.js';
import type { FleetModelTier, FleetRuntimeTarget } from '../lib/fleet-engine.js';
import { validateChannel } from '../shared/validators.js';

interface SpawnRouteDeps {
  spawner: Spawner;
  costTracker?: CostTracker;
  metrics: { errors: number };
  logger: {
    info(msg: string, meta?: Record<string, unknown>): void;
    error(msg: string, meta?: Record<string, unknown>): void;
  };
}

const VALID_BACKENDS = new Set(['ollama', 'lmstudio', 'claude', 'claude-cli', 'gemini', 'cloudflare', 'openai', 'groq', 'deepseek', 'xai', 'codex', 'aider', 'custom', 'cli:claude-code', 'cli:codex', 'cli:gemini', 'cli:groq', 'cli:grok']);


// ==========================================================================
// Fastify plugin (dual-export)
// ==========================================================================
export const spawnPlugin: FastifyPluginAsync<{ deps: SpawnRouteDeps }> = async (fastify, opts) => {
  const { metrics, logger, spawner, costTracker } = opts.deps;

  fastify.post('/spawn/preflight', async (request, reply) => {
    try {
      const body = (request.body as Record<string, unknown>) || {};
      const fallbacks = Array.isArray(body.fallbacks)
        ? (body.fallbacks.filter((value): value is FleetRuntimeTarget => !!value && typeof value === 'object'))
        : undefined;
      const budgetUsd = typeof body.budgetUsd === 'number'
        ? body.budgetUsd
        : typeof body.budgetUsd === 'string' && body.budgetUsd.trim()
          ? parseFloat(body.budgetUsd)
          : undefined;

      const preflight = await assessSpawnPreflight({
        backend: typeof body.backend === 'string' ? body.backend : undefined,
        model: typeof body.model === 'string' ? body.model : undefined,
        modelTier: typeof body.modelTier === 'string' ? body.modelTier as FleetModelTier : undefined,
        fallbacks,
        identity: typeof body.identity === 'string' ? body.identity : undefined,
        budgetUsd: Number.isFinite(budgetUsd) ? budgetUsd : undefined,
      }, { costTracker });

      return { success: true, ...preflight };
    } catch (error) {
      metrics.errors++;
      logger.error('spawn_preflight_error', { error: (error as Error).message });
      reply.code(500);
      return { error: 'internal server error' };
    }
  });

  // POST /spawn — Launch an AI agent
  fastify.post('/spawn', async (request, reply) => {
    try {
      const {
        backend,
        name,
        model,
        modelTier,
        identity,
        purpose,
        task,
        files,
        workdir,
        env,
        timeout,
        allowedTools,
        maxTokens,
        permissionMode,
        injectSquidHooks,
        tubeChannel,
        budgetUsd: rawBudgetUsd,
      } = request.body as any;

      if (!backend || typeof backend !== 'string') {
        reply.code(400); return {
          success: false,
          error: 'backend is required. Valid values: ollama, claude, claude-cli, gemini, cloudflare, codex, aider, custom',
          code: 'VALIDATION_ERROR',
        };
      }

      if (!VALID_BACKENDS.has(backend)) {
        reply.code(400); return {
          success: false,
          error: `Invalid backend "${backend}". Valid values: ${[...VALID_BACKENDS].join(', ')}`,
          code: 'VALIDATION_ERROR',
        };
      }

      if (!task || typeof task !== 'string' || !task.trim()) {
        reply.code(400); return {
          success: false,
          error: 'task is required and must be a non-empty string',
          code: 'VALIDATION_ERROR',
        };
      }

      if (typeof task === 'string' && task.length > 100000) {
        reply.code(400); return {
          success: false,
          error: 'task must not exceed 100000 characters',
          code: 'VALIDATION_ERROR',
        };
      }

      if (backend === 'custom' && /[;&|`$(){}!<>]/.test(task as string)) {
        reply.code(400); return {
          success: false,
          error: 'Custom backend task contains shell metacharacters. Use explicit arguments instead of shell syntax.',
          code: 'VALIDATION_ERROR',
        };
      }

      if (tubeChannel !== undefined && tubeChannel !== null) {
        const channelValidation = validateChannel(tubeChannel);
        if (!channelValidation.valid) {
          reply.code(400); return {
            success: false,
            error: `tubeChannel ${channelValidation.error}`,
            code: 'VALIDATION_ERROR',
          };
        }
      }

      // workdir is interpolated into the Coast Guard's OS-sandbox profile
      // (lib/coast-guard.ts buildSeatbeltProfile → `(subpath "<workdir>")`). A
      // quote/backslash/newline/NUL in it is an SBPL-injection vector (#339), so
      // reject it at the boundary — same posture as `task`'s metachar check.
      // We also require an absolute path: the sandbox confines an absolute root,
      // and a relative workdir is ambiguous against the daemon's cwd.
      if (workdir !== undefined && workdir !== null) {
        if (typeof workdir !== 'string' || !workdir.trim()) {
          reply.code(400); return {
            success: false,
            error: 'workdir must be a non-empty string',
            code: 'VALIDATION_ERROR',
          };
        }
        if (/["\\\n\r\0]/.test(workdir)) {
          reply.code(400); return {
            success: false,
            error: 'workdir contains an illegal character (quote, backslash, newline, or NUL). Provide a plain absolute path.',
            code: 'VALIDATION_ERROR',
          };
        }
        if (!workdir.startsWith('/')) {
          reply.code(400); return {
            success: false,
            error: 'workdir must be an absolute path (start with "/").',
            code: 'VALIDATION_ERROR',
          };
        }
      }

      const parsedBudgetUsd = typeof rawBudgetUsd === 'number'
        ? rawBudgetUsd
        : typeof rawBudgetUsd === 'string' && rawBudgetUsd.trim()
          ? parseFloat(rawBudgetUsd)
          : undefined;
      const preflight = await assessSpawnPreflight({
        backend,
        model,
        modelTier: typeof modelTier === 'string' ? modelTier as FleetModelTier : undefined,
        identity,
        budgetUsd: Number.isFinite(parsedBudgetUsd) ? parsedBudgetUsd : undefined,
      }, { costTracker });

      if (!preflight.launchReady) {
        reply.code(400);
        return {
          success: false,
          error: preflight.blockedReasons[0] || 'spawn preflight failed',
          code: 'PRECONDITION_FAILED',
          preflight,
        };
      }

      const spec: SpawnSpec = {
        backend: backend as SpawnSpec['backend'],
        task: task.trim(),
      };

      if (model && typeof model === 'string') spec.model = model;
      else if (preflight.attempts[0]?.model) spec.model = preflight.attempts[0].model;
      if (name && typeof name === 'string') spec.name = name;
      if (typeof modelTier === 'string') spec.modelTier = modelTier as FleetModelTier;
      else if (preflight.attempts[0]?.modelTier) spec.modelTier = preflight.attempts[0].modelTier as FleetModelTier;
      if (identity && typeof identity === 'string') spec.identity = identity;
      if (purpose && typeof purpose === 'string') spec.purpose = purpose;
      if (Array.isArray(files)) spec.files = files as string[];
      if (workdir && typeof workdir === 'string') spec.workdir = workdir;
      if (env && typeof env === 'object' && !Array.isArray(env)) spec.env = env as Record<string, string>;
      if (timeout && typeof timeout === 'number') spec.timeout = timeout;
      if (allowedTools && typeof allowedTools === 'string') spec.allowedTools = allowedTools;
      if (maxTokens && typeof maxTokens === 'number') spec.maxTokens = maxTokens;
      if (typeof tubeChannel === 'string') spec.tubeChannel = tubeChannel;
      // File-edit permission mode for the cli:claude-code backend. Only the three
      // CLI-recognised modes are accepted; anything else is ignored (the spawner
      // forwards it verbatim as --permission-mode, so the boundary validates it).
      if (permissionMode === 'default' || permissionMode === 'acceptEdits' || permissionMode === 'bypassPermissions') {
        spec.permissionMode = permissionMode;
      }
      // Giant Squid Harness opt-in (ADR-0091). Default false → backward-compatible:
      // an absent/false flag leaves the spawn byte-for-byte unchanged. When true,
      // the spawner's runClaudeCli (lib/spawner.ts) FIRST injects the pd-hook-*
      // tentacles into the workspace's .claude/settings.json, so a conjure-
      // dispatched vendor CLI runs UNDER PD coordination — its UserPromptSubmit /
      // PreToolUse / PostToolUse turns fire the lock gate + pheromone hooks inside
      // Claude Code's own loop (Claude Max Prime). The conjurer's Dispatch sets
      // this true (console DaemonClient::spawn with SpawnOpts::squid). codex /
      // gemini remain validate-then-add: their squid adapters throw, so the flag
      // is a harmless no-op for those backends until those adapters are written.
      if (injectSquidHooks === true) {
        spec.injectSquidHooks = true;
      }

      logger.info('spawn_start', {
        backend,
        model: spec.model || null,
        identity: spec.identity || null,
        purpose: spec.purpose || null,
      });

      const result = await spawner.spawn(spec);

      logger.info('spawn_complete', {
        agentId: result.agentId,
        backend: result.backend,
        status: result.status,
      });

      return { success: true, ...result };
    } catch (error) {
      metrics.errors++;
      logger.error('spawn_error', { error: (error as Error).message });
      reply.code(500); return { error: 'internal server error' };
    }
  });

  // GET /spawn — List active spawned agents
  fastify.get('/spawn', async (request, reply) => {
    try {
      const agents = spawner.list();
      return {
        success: true,
        agents,
        count: agents.length,
      };
    } catch (error) {
      metrics.errors++;
      logger.error('spawn_list_error', { error: (error as Error).message });
      reply.code(500); return { error: 'internal server error' };
    }
  });

  // DELETE /spawn/:id — Kill a spawned agent
  fastify.delete('/spawn/:id', async (request, reply) => {
    try {
      const id = String((request.params as any).id);

      spawner.kill(id);

      logger.info('spawn_kill', { agentId: id });

      return {
        success: true,
        agentId: id,
        message: `Agent ${id} killed`,
      };
    } catch (error) {
      metrics.errors++;
      logger.error('spawn_kill_error', { error: (error as Error).message });
      reply.code(500); return { error: 'internal server error' };
    }
  });
};